requirements.txt
arrow==1.3.0
asgiref==3.8.1
bcrypt==4.2.0
binaryornot==0.4.4
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cryptography==43.0.0
Django==5.1
django-browser-reload==1.13.0
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-restframework==0.0.1
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
secure-smtplib==0.1.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
